# fix(ngcc): handle entry-points within container folders