--- a/packages/compiler-cli/ngcc/src/entry_point_finder/utils.ts
+++ b/packages/compiler-cli/ngcc/src/entry_point_finder/utils.ts
@@ -78,3 +78,18 @@ function removeContainedPaths(value: AbsoluteFsPath, index: number, array: Absol
   }
   return true;
 }
+
+/**
+ * Run a task and track how long it takes.
+ *
+ * @param task The task whose duration we are tracking
+ * @param log The function to call with the duration of the task
+ * @returns The result of calling `task`.
+ */
+export function trackDuration<T = void>(task: () => T, log: (duration: number) => void): T {
+  const startTime = Date.now();
+  const result = task();
+  const duration = Math.round((Date.now() - startTime) / 100) / 10;
+  log(duration);
+  return result;
+}