--- a/packages/compiler-cli/ngcc/src/entry_point_finder/utils.ts
+++ b/packages/compiler-cli/ngcc/src/entry_point_finder/utils.ts
@@ -86,7 +86,8 @@ function removeContainedPaths(value: AbsoluteFsPath, index: number, array: Absol
  * @param log The function to call with the duration of the task
  * @returns The result of calling `task`.
  */
-export function trackDuration<T = void>(task: () => T, log: (duration: number) => void): T {
+export function trackDuration<T = void>(
+    task: () => T extends Promise<unknown>? never : T, log: (duration: number) => void): T {
   const startTime = Date.now();
   const result = task();
   const duration = Math.round((Date.now() - startTime) / 100) / 10;