--- a/packages/compiler-cli/ngcc/test/writing/new_entry_point_file_writer_spec.ts
+++ b/packages/compiler-cli/ngcc/test/writing/new_entry_point_file_writer_spec.ts
@@ -9,7 +9,7 @@ import {FileSystem, absoluteFrom, getFileSystem} from '../../../src/ngtsc/file_s
 import {runInEachFileSystem} from '../../../src/ngtsc/file_system/testing';
 import {loadTestFiles} from '../../../test/helpers';
 import {NgccConfiguration} from '../../src/packages/configuration';
-import {EntryPoint, EntryPointFormat, EntryPointJsonProperty, INVALID_ENTRY_POINT, NO_ENTRY_POINT, getEntryPointInfo} from '../../src/packages/entry_point';
+import {EntryPoint, EntryPointFormat, EntryPointJsonProperty, INCOMPATIBLE_ENTRY_POINT, NO_ENTRY_POINT, getEntryPointInfo} from '../../src/packages/entry_point';
 import {EntryPointBundle, makeEntryPointBundle} from '../../src/packages/entry_point_bundle';
 import {FileWriter} from '../../src/writing/file_writer';
 import {NewEntryPointFileWriter} from '../../src/writing/new_entry_point_file_writer';
@@ -108,7 +108,7 @@ runInEachFileSystem(() => {
         const config = new NgccConfiguration(fs, _('/'));
         const result = getEntryPointInfo(
             fs, config, logger, _('/node_modules/test'), _('/node_modules/test')) !;
-        if (result === NO_ENTRY_POINT || result === INVALID_ENTRY_POINT) {
+        if (result === NO_ENTRY_POINT || result === INCOMPATIBLE_ENTRY_POINT) {
           return fail(`Expected an entry point but got ${result}`);
         }
         entryPoint = result;
@@ -248,7 +248,7 @@ runInEachFileSystem(() => {
         const config = new NgccConfiguration(fs, _('/'));
         const result = getEntryPointInfo(
             fs, config, logger, _('/node_modules/test'), _('/node_modules/test/a')) !;
-        if (result === NO_ENTRY_POINT || result === INVALID_ENTRY_POINT) {
+        if (result === NO_ENTRY_POINT || result === INCOMPATIBLE_ENTRY_POINT) {
           return fail(`Expected an entry point but got ${result}`);
         }
         entryPoint = result;
@@ -377,7 +377,7 @@ runInEachFileSystem(() => {
         const config = new NgccConfiguration(fs, _('/'));
         const result = getEntryPointInfo(
             fs, config, new MockLogger(), _('/node_modules/test'), _('/node_modules/test/b')) !;
-        if (result === NO_ENTRY_POINT || result === INVALID_ENTRY_POINT) {
+        if (result === NO_ENTRY_POINT || result === INCOMPATIBLE_ENTRY_POINT) {
           return fail(`Expected an entry point but got ${result}`);
         }
         entryPoint = result;