… packages/compiler-cli/ngcc/test/packages/entry_point_spec.ts