@petebacondarwin 2020-03-30
I'm going to run a few performance measurements against master this morning just to see what impact this has.