@petebacondarwin 2020-04-01
The perf tests are in (at least for my OS/X machine). I see no statistically significant difference between the `master` branch and this branch for running `ngcc`against the AIO `node_modules`.

**Time to walk directories**

run | master | PR
---- | -------- | ---
1    | 2.3 secs      | 1.0 secs
2    | 1.0 secs       | 0.9 secs
3    | 1.0 secs       | 1.3 secs